feat(examples-advanced): split out & basic cleanup of save & load
